+++ src/App.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
 import logo from './logo.svg';
 import MainPage from './Pages/MainPage'
+import Header from './Components/Header/Header';
 import './App.css';
 
 function App() {
   return (
+
     <MainPage />
   );
 }
