+++ src/Components/Directions/Directions.css
@@ -1,0 +1,162 @@
+/*--------------------------- Directions ------------------------*/
+
+.directions {
+    margin-top: 300px;
+  }
+  
+  .directions_header {
+    text-align: center;
+    margin-bottom: 145px;
+  }
+  
+  .directions_header h2 {
+    font-size: var(--section-header-font-size);
+  }
+  
+  .directions_items {
+    display: flex;
+    justify-content: space-between;
+  }
+  
+  .directions_item {
+    width: 406px;
+    height: 460px;
+    border-radius: 40px;
+    text-align: center;
+  }
+  
+  .directions_item-course {
+    text-align: right;
+    color: rgba(255, 255, 255, 0.80);
+    font-size: 18px;
+    font-style: normal;
+    font-weight: 400;
+    line-height: normal;
+    padding-right: 40px;
+    margin-top: 48px;
+  }
+  
+  
+  .directions_item-code {
+    font-family: var(--main-semibold-font-family);
+    font-size: 70px;
+    font-style: normal;
+    font-weight: 600;
+    line-height: normal;
+    margin-top: 73px;
+  }
+  
+  
+  .directions_item-name {
+    margin-top: 35px;
+    font-family: var(--main-font-family);
+    font-size: 22px;
+    margin-bottom: 100px;
+    font-weight:bolder;
+    line-height: normal;
+    height: 47px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  
+  .gradient1-layer1, .gradient1-layer2, .gradient2-layer1, .gradient2-layer2, .gradient2-layer3,.gradient2-layer4,  .gradient3-layer1, .gradient3-layer2, .gradient3-layer3 {
+    width: 406px;
+    position: absolute;
+    height: 460px;
+    border-radius: 40px;
+  }
+  .gradient1-layer1 {
+    z-index: -2;
+    background: radial-gradient(100.61% 123.38% at 40.57% -23.81%, rgba(180, 81, 242, 0.84) 0%, rgba(253, 243, 243, 0.10) 100%);
+    -webkit-background: radial-gradient(100.61% 123.38% at 40.57% -23.81%, rgba(180, 81, 242, 0.84) 0%, rgba(253, 243, 243, 0.10) 100%);
+    -moz-background: radial-gradient(100.61% 123.38% at 40.57% -23.81%, rgba(180, 81, 242, 0.84) 0%, rgba(253, 243, 243, 0.10) 100%);
+    -o-background: radial-gradient(100.61% 123.38% at 40.57% -23.81%, rgba(180, 81, 242, 0.84) 0%, rgba(253, 243, 243, 0.10) 100%);
+    background-blend-mode: screen;
+    backdrop-filter: blur(25px);
+    opacity: 0.5;
+  }
+  
+  .gradient1-layer2 {
+    z-index: -1;
+    background: radial-gradient(157.07% 93.98% at 14.41% 119.18%, rgba(3, 11, 255, 0.289) 0.73%, rgba(253, 243, 243, 0.10) 100%);
+    -webkit-background: radial-gradient(157.07% 93.98% at 14.41% 119.18%, rgba(3, 11, 255, 0.289) 0.73%, rgba(253, 243, 243, 0.10) 100%);
+    -moz-background: radial-gradient(157.07% 93.98% at 14.41% 119.18%, rgba(3, 11, 255, 0.289) 0.73%, rgba(253, 243, 243, 0.10) 100%);
+    -o-background: radial-gradient(157.07% 93.98% at 14.41% 119.18%, rgba(3, 11, 255, 0.289) 0.73%, rgba(253, 243, 243, 0.10) 100%);
+    background-blend-mode: screen;
+    mix-blend-mode: color-dodge;
+    backdrop-filter: blur(25px);
+    opacity: 0.6;
+  }
+  
+  .gradient2-layer1 {
+    z-index: -2;
+    background: radial-gradient(90.64% 144.06% at 173.89% -9.68%, rgba(163, 62, 179, 0.50) 0%, rgba(253, 243, 243, 0.062) 98.85%);
+    -webkit-background: radial-gradient(90.64% 144.06% at 173.89% -9.68%, rgba(163, 62, 179, 0.50) 0%, rgba(253, 243, 243, 0.062) 98.85%);
+    -moz-background: radial-gradient(90.64% 144.06% at 173.89% -9.68%, rgba(163, 62, 179, 0.50) 0%, rgba(253, 243, 243, 0.062) 98.85%);
+    -o-background: radial-gradient(90.64% 144.06% at 173.89% -9.68%, rgba(163, 62, 179, 0.50) 0%, rgba(253, 243, 243, 0.062) 98.85%);
+    backdrop-filter: blur(50px);
+    opacity: 0.7;
+  }
+  
+  .gradient2-layer2 {
+    z-index: -1;
+    background: radial-gradient(45.86% 52.97% at 95.16% 3.93%, rgba(180, 81, 242, 0.84) 0%, rgba(253, 243, 243, 0.10) 100%);
+    -webkit-background: radial-gradient(45.86% 52.97% at 95.16% 3.93%, rgba(180, 81, 242, 0.84) 0%, rgba(253, 243, 243, 0.10) 100%);
+    -moz-background: radial-gradient(45.86% 52.97% at 95.16% 3.93%, rgba(180, 81, 242, 0.84) 0%, rgba(253, 243, 243, 0.10) 100%);
+    -o-background: radial-gradient(45.86% 52.97% at 95.16% 3.93%, rgba(180, 81, 242, 0.84) 0%, rgba(253, 243, 243, 0.10) 100%);
+    mix-blend-mode: color-dodge;
+    backdrop-filter: blur(25px);
+    opacity: 0.4;
+  }
+  
+  .gradient2-layer3 {
+    z-index: -2;
+    background: radial-gradient(43.86% 34.97% at 6.16% 53.93%, rgba(143, 69, 240, 0.717) 0%, rgba(253, 243, 243, 0.10) 100%);
+    -webkit-background: radial-gradient(43.86% 34.97% at 6.16% 53.93%, rgba(143, 69, 240, 0.717) 0%, rgba(253, 243, 243, 0.10) 100%);
+    -moz-background: radial-gradient(43.86% 34.97% at 6.16% 53.93%, rgba(143, 69, 240, 0.717) 0%, rgba(253, 243, 243, 0.10) 100%);
+    -o-background: radial-gradient(43.86% 34.97% at 6.16% 53.93%, rgba(143, 69, 240, 0.717) 0%, rgba(253, 243, 243, 0.10) 100%);
+    background-blend-mode: screen;
+    mix-blend-mode: color-dodge;
+    backdrop-filter: blur(25px);
+    opacity: 0.4; 
+  }
+  
+  .gradient2-layer4 {
+    z-index: -3;
+    background:#A972EF;
+    opacity: 0.1; 
+  }
+  
+  .gradient3-layer1 {
+    z-index: -3;
+    background: radial-gradient(90.64% 144.06% at 173.89% -9.68%, rgba(180, 66, 250, 0.84) 0%, rgba(126, 59, 157, 0.195) 98.85%);
+    -webkit-background: radial-gradient(90.64% 144.06% at 173.89% -9.68%, rgba(180, 66, 250, 0.84) 0%, rgba(126, 59, 157, 0.195) 98.85%);
+    -moz-background: radial-gradient(90.64% 144.06% at 173.89% -9.68%, rgba(180, 66, 250, 0.84) 0%, rgba(126, 59, 157, 0.195) 98.85%);
+    -o-background: radial-gradient(90.64% 144.06% at 173.89% -9.68%, rgba(180, 66, 250, 0.84) 0%, rgba(126, 59, 157, 0.195) 98.85%);
+    background-blend-mode: screen;
+    backdrop-filter: blur(50px);
+    opacity: 0.4;
+  }
+  
+  .gradient3-layer2 {
+    z-index: -2;
+    background: radial-gradient(60.86% 60.97% at 3.16% 3.93%, rgba(180, 66, 250, 0.84) 0%, rgba(253, 243, 243, 0.10) 100%);
+    -webkit-background: radial-gradient(60.86% 60.97% at 3.16% 3.93%, rgba(180, 66, 250, 0.84) 0%, rgba(253, 243, 243, 0.10) 100%);
+    -moz-background: radial-gradient(60.86% 60.97% at 3.16% 3.93%, rgba(180, 66, 250, 0.84) 0%, rgba(253, 243, 243, 0.10) 100%);
+    -o-background: radial-gradient(60.86% 60.97% at 3.16% 3.93%, rgba(180, 66, 250, 0.84) 0%, rgba(253, 243, 243, 0.10) 100%);
+    background-blend-mode: screen;
+    mix-blend-mode: color-dodge;
+    backdrop-filter: blur(25px);
+    opacity: 0.4;
+  }
+  
+  .gradient3-layer3 {
+    z-index: -1;
+    background: radial-gradient(157.07% 93.98% at 14.41% 119.18%, rgba(3, 11, 255, 0.289) 0.73%, rgba(253, 243, 243, 0.10) 100%);
+    -webkit-background: radial-gradient(157.07% 93.98% at 14.41% 119.18%, rgba(3, 11, 255, 0.289) 0.73%, rgba(253, 243, 243, 0.10) 100%);
+    -moz-background: radial-gradient(157.07% 93.98% at 14.41% 119.18%, rgba(3, 11, 255, 0.289) 0.73%, rgba(253, 243, 243, 0.10) 100%);
+    -o-background: radial-gradient(157.07% 93.98% at 14.41% 119.18%, rgba(3, 11, 255, 0.289) 0.73%, rgba(253, 243, 243, 0.10) 100%);
+    backdrop-filter: blur(25px);
+    opacity: 0.4;
+  }
