+++ src/Components/Footer/Footer.css
@@ -1,0 +1,52 @@
+.footer {
+  margin-top: 300px;
+}
+
+.footer_header {
+  align-items: left;
+}
+
+.footer_header h2 {
+  font-size: var(--section-header-font-size);
+}
+
+.footer_text {
+  font-size: 30px;
+  line-height: 40px;
+  letter-spacing: 0.03em;
+  text-align: left;
+}
+
+.footer_mail {
+  margin-top: 52px;
+  height: 40px;
+  display: flex;
+  flex-direction: row;
+  column-gap: 70px;
+  align-items: center;
+  row-gap: 71px;
+}
+
+.footer_location {
+  margin-top: 71px;
+  width: 1216px;
+  height: 88px;
+
+  display: flex;
+  flex-direction: row;
+  column-gap: 70px;
+  align-items: center;
+}
+
+.footer_icon {
+  width: 50px;
+  height: 50px;
+}
+
+.footer_images {
+  margin-top: 113px;
+  display: flex;
+  column-gap: 58px;
+  align-items: center;
+  padding-bottom: 90px;
+}
