+++ src/Components/Footer/Footer.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 
-import '../App.css';
+import './Footer.css';
 
 
 const Footer: FC = () => {
