+++ src/Components/Main/Main.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from 'react';
 
-import '../App.css';
+import './Main.css';
 
 
 const Main: FC = () => {
